+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Log level to args.
+
+### Fixed
+- Logger not logging to stdout.
 
 ## [0.1.0-beta.1] - 2020-03-10
 ### Added